requirements.txt
flask
pymysql
slackclient>=2.0.0
